+++ Flosion/convolver.h
@@ -40,15 +40,15 @@ namespace musical {
 
 		void renderChunk(Sample* buffer, ConvolverState* state) override {
 			state->length = std::min(std::max(0u, (unsigned)(function_length.getValue(state) * SFREQ)), max_convolution_length);
-			for (int i = 0; i < state->length; i++){
+			for (unsigned int i = 0; i < state->length; i++){
 				state->position = i;
 				state->conv[i] = function.getValue(state);
 			}
 
 			state->queue.advance(CHUNK_SIZE, input, state);
-			for (int i = 0; i < CHUNK_SIZE; i++){
+			for (size_t i = 0; i < CHUNK_SIZE; i++){
 				Sample s = Sample(0, 0);
-				for (int j = 0; j < state->length; j++){
+				for (size_t j = 0; j < state->length; j++){
 					s += state->queue[i + j] * state->conv[j];
 				}
 				buffer[i] = s / (float)SFREQ;
